+++ bash_test.go
@@ -1,9 +1,7 @@
 package main
 
 import (
-	"bytes"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
 	"testing"
@@ -29,13 +27,13 @@ func TestBashApplier_ApplyHeader(t *testing.T) {
 		fileName := f
 		t.Run(fileName, func(t *testing.T) {
 			testfile := filepath.Join(tmpDir, fileName)
-			copyBashFile(t, "./testdata/"+fileName, testfile)
+			copyFile(t, "./testdata/"+fileName, testfile)
 
 			err := d.ApplyHeader(testfile, &tc)
 			if err != nil {
 				t.Fatalf("failed to apply header to %s: %+v", testfile, err)
 			}
-			compareBashFiles(t, testfile, "./testdata/"+fileName+".golden")
+			compareFiles(t, testfile, "./testdata/"+fileName+".golden")
 		})
 	}
 }
@@ -91,37 +89,5 @@ func newBashTagContext(t *testing.T) TagContext {
 	return TagContext{
 		templateFiles: TemplateFiles{shTemplateFile: templateFile},
 		templatePath:  "./testdata/templates/",
-	}
-}
-
-func copyBashFile(t *testing.T, src, dst string) {
-	t.Helper()
-	input, err := os.ReadFile(src)
-	if err != nil {
-		t.Fatalf("failed to read %s: %+v", src, err)
-	}
-	err = os.WriteFile(dst, input, 0777)
-	if err != nil {
-		t.Fatalf("failed to write %s: %+v", dst, err)
-	}
-}
-
-func compareBashFiles(t *testing.T, a, b string) {
-	contentA, err := os.ReadFile(a)
-	if err != nil {
-		t.Fatalf("failed to read file %s: %+v", a, err)
-	}
-	contentB, err := os.ReadFile(b)
-	if err != nil {
-		t.Fatalf("failed to read golden file %s: %+v", b, err)
-	}
-	if !bytes.Equal(contentA, contentB) {
-		if _, err := exec.LookPath("git"); err == nil {
-			diff := exec.Command("git", "diff", "--no-index", a, b)
-			stdoutStderr, _ := diff.CombinedOutput()
-			t.Fatalf("%s\n", stdoutStderr)
-		} else {
-			t.Fatalf("content of %s did not match %s.\nContent A:\n%s\n\nContent B:\n%s\n", a, b, contentA, contentB)
-		}
 	}
 }
